+++ examples/imgui_backend_sea.py
@@ -9,7 +9,7 @@ import wgpu
 import time
 import numpy as np
 from imgui_bundle import imgui
-from wgpu.utils.imgui_backend import ImguiWgpuBackend
+from wgpu.utils.imgui import ImguiWgpuBackend
 
 # Create a canvas to render to
 canvas = WgpuCanvas(title="imgui_sea", size=(800, 450), max_fps=60)
