+++ examples/imgui_renderer_sea.py
@@ -9,7 +9,7 @@ import wgpu
 import time
 import numpy as np
 from imgui_bundle import imgui
-from wgpu.utils.imgui_renderer import ImguiRenderer
+from wgpu.utils.imgui import ImguiRenderer
 
 # Create a canvas to render to
 canvas = WgpuCanvas(title="imgui_sea", size=(800, 450), max_fps=60)
